It seems to be a bug when the step defined for offset is 1. The bug is getting the offset, retrieves the same as before.
This happens for button_up and button_left. They doesn't applies any change.

For any other step (such as 10) works like a charm.

Probably is because to  something else which will be analised in the future.

The default was left in 10 instead 1 . The same default is when text leaved in
blanc